[vk] Move PhysicalDevice to its own file
